+++ tests/Registration/ServiceTest.php
@@ -1,0 +1,11 @@
+<?php
+
+use PHPUnit\Framework\TestCase;
+
+class ServiceTest extends TestCase
+{
+    public function test_post_type_returns_defaults()
+    {
+        //
+    }
+}
